# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ResumeProfile = { [key: string]: unknown };

export type ResumeExperienceItem = { [key: string]: unknown };

export type ResumeEducationItem = { [key: string]: unknown };

export type ResumeSkillsItem = { [key: string]: unknown };

export type ResumeProjectsItem = { [key: string]: unknown };

export type ResumeCertificationsItem = { [key: string]: unknown };

export interface Resume {
  id: number;
  title: string;
  template: string;
  profile: ResumeProfile;
  experience: ResumeExperienceItem[];
  education: ResumeEducationItem[];
  skills: ResumeSkillsItem[];
  projects: ResumeProjectsItem[];
  certifications: ResumeCertificationsItem[];
  atsScore?: number | null;
  targetRole?: string | null;
  targetIndustry?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateResumeBody {
  title: string;
  template?: string;
}

export type UpdateResumeBodyProfile = { [key: string]: unknown };

export type UpdateResumeBodyExperienceItem = { [key: string]: unknown };

export type UpdateResumeBodyEducationItem = { [key: string]: unknown };

export type UpdateResumeBodySkillsItem = { [key: string]: unknown };

export type UpdateResumeBodyProjectsItem = { [key: string]: unknown };

export type UpdateResumeBodyCertificationsItem = { [key: string]: unknown };

export interface UpdateResumeBody {
  title?: string;
  template?: string;
  profile?: UpdateResumeBodyProfile;
  experience?: UpdateResumeBodyExperienceItem[];
  education?: UpdateResumeBodyEducationItem[];
  skills?: UpdateResumeBodySkillsItem[];
  projects?: UpdateResumeBodyProjectsItem[];
  certifications?: UpdateResumeBodyCertificationsItem[];
  targetRole?: string;
  targetIndustry?: string;
}

/**
 * Current section content
 */
export type ImproveSectionBodyContent = { [key: string]: unknown };

export type ImproveSectionBodyPreviousVersion = {
  [key: string]: unknown;
} | null;

export interface ImproveSectionBody {
  resumeId?: number;
  /** Section name (profile, experience, education, skills, projects) */
  section: string;
  /** Current section content */
  content: ImproveSectionBodyContent;
  targetRole?: string | null;
  targetIndustry?: string | null;
  yearsOfExperience?: number | null;
  /** startup, corporate, faang, consulting */
  companyType?: string | null;
  /** concise, detailed, technical, business */
  tone?: string | null;
  /** Specific metrics/achievements to include */
  metrics?: string | null;
  previousVersion?: ImproveSectionBodyPreviousVersion;
}

export type AtsScoreBodyResumeContent = { [key: string]: unknown };

export interface AtsScoreBody {
  resumeId?: number;
  jobDescription?: string | null;
  resumeContent: AtsScoreBodyResumeContent;
}

export type AtsScoreResultBreakdown = {
  keywords: number;
  structure: number;
  contentStrength: number;
  readability: number;
  measurableAchievements: number;
};

export type AtsScoreResultSectionScores = { [key: string]: number };

export interface AtsScoreResult {
  totalScore: number;
  breakdown: AtsScoreResultBreakdown;
  missingKeywords: string[];
  suggestions: string[];
  sectionScores: AtsScoreResultSectionScores;
}

export type JobTargetingBodyResumeContent = { [key: string]: unknown };

export interface JobTargetingBody {
  resumeId?: number | null;
  resumeContent: JobTargetingBodyResumeContent;
  jobDescription: string;
  confirmOptimize?: boolean;
}

export type ResumeChatBodyResumeContent = { [key: string]: unknown } | null;

export type ResumeChatBodyConversationHistoryItem = {
  role: string;
  content: string;
};

export interface ResumeChatBody {
  resumeId?: number | null;
  resumeContent?: ResumeChatBodyResumeContent;
  message: string;
  conversationHistory?: ResumeChatBodyConversationHistoryItem[];
}

export interface ErrorResponse {
  error: string;
}