# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resume OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/improve-section:
    post:
      operationId: improveSection
      tags: [ai]
      summary: AI-improve a resume section (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveSectionBody"
      responses:
        "200":
          description: SSE stream of AI improvements
          content:
            text/event-stream: {}
  /ai/ats-score:
    post:
      operationId: getAtsScore
      tags: [ai]
      summary: Get ATS score for a resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtsScoreBody"
      responses:
        "200":
          description: ATS score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtsScoreResult"
  /ai/job-targeting:
    post:
      operationId: targetJob
      tags: [ai]
      summary: Analyze resume against job description and suggest improvements (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobTargetingBody"
      responses:
        "200":
          description: SSE stream of job targeting suggestions
          content:
            text/event-stream: {}
  /ai/chat:
    post:
      operationId: resumeChat
      tags: [ai]
      summary: Global AI chat about the resume (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeChatBody"
      responses:
        "200":
          description: SSE stream of AI chat response
          content:
            text/event-stream: {}
components:
  schemas:
...
    ImproveSectionBody:
      type: object
      properties:
        resumeId:
          type: integer
        section:
          type: string
          description: Section name (profile, experience, education, skills, projects)
        content:
          type: object
          additionalProperties: true
          description: Current section content
        targetRole:
          type: string
          nullable: true
        targetIndustry:
          type: string
          nullable: true
        yearsOfExperience:
          type: integer
          nullable: true
        companyType:
          type: string
          nullable: true
          description: startup, corporate, faang, consulting
        tone:
          type: string
          nullable: true
          description: concise, detailed, technical, business
        metrics:
          type: string
          nullable: true
          description: Specific metrics/achievements to include
        previousVersion:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - section
        - content
    AtsScoreBody:
      type: object
      properties:
        resumeId:
          type: integer
        jobDescription:
          type: string
          nullable: true
        resumeContent:
          type: object
          additionalProperties: true
      required:
        - resumeContent
    AtsScoreResult:
      type: object
      properties:
        totalScore:
          type: integer
        breakdown:
          type: object
          properties:
            keywords:
              type: integer
            structure:
              type: integer
            contentStrength:
              type: integer
            readability:
              type: integer
            measurableAchievements:
              type: integer
          required:
            - keywords
            - structure
            - contentStrength
            - readability
            - measurableAchievements
        missingKeywords:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        sectionScores:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalScore
        - breakdown
        - missingKeywords
        - suggestions
        - sectionScores
    JobTargetingBody:
      type: object
      properties:
        resumeId:
          type: integer
          nullable: true
        resumeContent:
          type: object
          additionalProperties: true
        jobDescription:
          type: string
        confirmOptimize:
          type: boolean
      required:
        - resumeContent
        - jobDescription
    ResumeChatBody:
      type: object
      properties:
        resumeId:
          type: integer
          nullable: true
        resumeContent:
          type: object
          additionalProperties: true
          nullable: true
        message:
          type: string
        conversationHistory:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
      required:
        - message
...