# lib/api-zod/src/generated/types/atsScoreBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AtsScoreBodyResumeContent } from "./atsScoreBodyResumeContent";

export interface AtsScoreBody {
  resumeId?: number;
  jobDescription?: string | null;
  resumeContent: AtsScoreBodyResumeContent;
}