# lib/api-zod/src/generated/types/improveSectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImproveSectionBodyContent } from "./improveSectionBodyContent";
import type { ImproveSectionBodyPreviousVersion } from "./improveSectionBodyPreviousVersion";

export interface ImproveSectionBody {
  resumeId?: number;
  /** Section name (profile, experience, education, skills, projects) */
  section: string;
  /** Current section content */
  content: ImproveSectionBodyContent;
  targetRole?: string | null;
  targetIndustry?: string | null;
  yearsOfExperience?: number | null;
  /** startup, corporate, faang, consulting */
  companyType?: string | null;
  /** concise, detailed, technical, business */
  tone?: string | null;
  /** Specific metrics/achievements to include */
  metrics?: string | null;
  previousVersion?: ImproveSectionBodyPreviousVersion;
}