# lib/api-zod/src/generated/types/resumeChatBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeChatBodyConversationHistoryItem } from "./resumeChatBodyConversationHistoryItem";
import type { ResumeChatBodyResumeContent } from "./resumeChatBodyResumeContent";

export interface ResumeChatBody {
  resumeId?: number | null;
  resumeContent?: ResumeChatBodyResumeContent;
  message: string;
  conversationHistory?: ResumeChatBodyConversationHistoryItem[];
}